feat: pricing

+++ Assets/Economic/Interfaces/IWallet.cs
@@ -1,13 +1,15 @@
+using System.Numerics;
+
 namespace Factory
 {
 	public interface IWallet
 	{
-		public float GetBalance(CurrencyType type);
-		public void CurrencyTransfer(CurrencyType type, float value);
-		public void CurrencyTransfer(CurrencyType type, float value, out float walletBalance);
-		public void CurrencySet(CurrencyType type, float value);
+		public long GetBalance(CurrencyType type);
+		public void CurrencyTransfer(CurrencyType type, long value);
+		public void CurrencyTransfer(CurrencyType type, long value, out long walletBalance);
+		public void CurrencySet(CurrencyType type, long value);
 		public void CurrencySetDefault(CurrencyType type);
-		public void CurrencySetAll(float value);
+		public void CurrencySetAll(long value);
 		public void CurrencySetDefaultAll();
 	}
 }
